+++ .github/workflows/tests.yaml
@@ -88,7 +88,7 @@ jobs:
               #
               run: |
                   ./scripts/run_in_build_env.sh \
-                    "./scripts/py_matter_idl/matter/idl/zapxml_parser.py \
+                    "matter-zapxml-parser \
                       --no-print \
                       --log-level info \
                       src/app/zap-templates/zcl/data-model/chip/global-attributes.xml \
