+++ docs/guides/matter_idl_tooling.md
@@ -20,12 +20,12 @@ information available in [data_model/README.md](../../data_model/README.md).
 NOTE: scraper is a work in progress, XML data may be incomplete or have errors
 still.
 
-The script `./scripts/py_matter_idl/matter/idl/data_model_xml_parser.py` has the
-ability to parse one or more CSA data model XML files and output their content
-in `.matter` format. For example:
+The script `matter-data-model-xml-parser` has the ability to parse one or more
+CSA data model XML files and output their content in `.matter` format. For
+example:
 
 ```sh
-./scripts/py_matter_idl/matter/idl/data_model_xml_parser.py data_model/clusters/BooleanState.xml
+matter-data-model-xml-parser data_model/clusters/BooleanState.xml
 ```
 
 The tool supports several options that are useful for development:
@@ -54,7 +54,7 @@ specification:
 As such one can run compares such as:
 
 ```sh
-./scripts/py_matter_idl/matter/idl/data_model_xml_parser.py         \
+matter-data-model-xml-parser \
      -o out/spec.matter                                             \
      --compare-output out/sdk.matter                                \
      --compare src/controller/data_model/controller-clusters.matter \
