+++ scripts/py_matter_idl/matter/idl/lint/__init__.py
@@ -12,6 +12,43 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import logging
+
+import click
+
 from .lint_rules_parser import CreateParser
 
 __all__ = ['CreateParser']
+
+# Supported log levels, mapping string values required for argument
+# parsing into logging constants
+__LOG_LEVELS__ = {
+    'debug': logging.DEBUG,
+    'info': logging.INFO,
+    'warn': logging.WARN,
+    'fatal': logging.FATAL,
+}
+
+
+@click.command()
+@click.option(
+    '--log-level',
+    default='INFO',
+    type=click.Choice(list(__LOG_LEVELS__.keys()), case_sensitive=False),
+    help='Determines the verbosity of script output.')
+@click.argument('filename')
+def main(log_level, filename=None):
+    # The IDL lint parser is generally not intended to be run as a stand-alone binary.
+    # The ability to run is for debug and to print out the parsed AST.
+
+    logging.basicConfig(
+        level=__LOG_LEVELS__[log_level],
+        format='%(asctime)s %(levelname)-7s %(message)s',
+    )
+
+    logging.info("Starting to parse ...")
+    data = CreateParser(filename).parse()
+    logging.info("Parse completed")
+
+    logging.info("Data:")
+    logging.info("%r" % data)
