+++ scripts/py_matter_idl/matter/idl/lint/lint_rules_parser.py
@@ -356,40 +356,3 @@ def CreateParser(file_name: str):
     """
     return Parser(
         Lark.open('lint_rules_grammar.lark', rel_to=__file__, parser='lalr', propagate_positions=True, maybe_placeholders=True), file_name=file_name)
-
-
-if __name__ == '__main__':
-    # This Parser is generally not intended to be run as a stand-alone binary.
-    # The ability to run is for debug and to print out the parsed AST.
-    import click
-
-    # Supported log levels, mapping string values required for argument
-    # parsing into logging constants
-    __LOG_LEVELS__ = {
-        'debug': logging.DEBUG,
-        'info': logging.INFO,
-        'warn': logging.WARN,
-        'fatal': logging.FATAL,
-    }
-
-    @click.command()
-    @click.option(
-        '--log-level',
-        default='INFO',
-        type=click.Choice(list(__LOG_LEVELS__.keys()), case_sensitive=False),
-        help='Determines the verbosity of script output.')
-    @click.argument('filename')
-    def main(log_level, filename=None):
-        logging.basicConfig(
-            level=__LOG_LEVELS__[log_level],
-            format='%(asctime)s %(levelname)-7s %(message)s',
-        )
-
-        logging.info("Starting to parse ...")
-        data = CreateParser(filename).parse()
-        logging.info("Parse completed")
-
-        logging.info("Data:")
-        logging.info("%r" % data)
-
-    main(auto_envvar_prefix='CHIP')
