+++ scripts/py_matter_idl/matter/idl/matter_idl_parser.py
@@ -3,10 +3,12 @@
 import dataclasses
 import functools
 import logging
+import pprint
 import sys
 from pathlib import Path
 from typing import Dict, List, Optional
 
+import click
 from lark import Lark
 from lark.lexer import Token
 from lark.visitors import Transformer, v_args
@@ -711,40 +713,35 @@ def CreateParser(skip_meta: bool = False, merge_globals=True):
     return ParserWithLines(skip_meta, merge_globals)
 
 
-if __name__ == '__main__':
-    # This Parser is generally not intended to be run as a stand-alone binary.
+# Supported log levels, mapping string values required for argument
+# parsing into logging constants
+__LOG_LEVELS__ = {
+    'debug': logging.DEBUG,
+    'info': logging.INFO,
+    'warn': logging.WARN,
+    'fatal': logging.FATAL,
+}
+
+
+@click.command()
+@click.option(
+    '--log-level',
+    default='INFO',
+    type=click.Choice(list(__LOG_LEVELS__.keys()), case_sensitive=False),
+    help='Determines the verbosity of script output.')
+@click.argument('filename')
+def main(log_level, filename=None):
+    # The IDL parser is generally not intended to be run as a stand-alone binary.
     # The ability to run is for debug and to print out the parsed AST.
-    import pprint
 
-    import click
+    logging.basicConfig(
+        level=__LOG_LEVELS__[log_level],
+        format='%(asctime)s %(levelname)-7s %(message)s',
+    )
 
-    # Supported log levels, mapping string values required for argument
-    # parsing into logging constants
-    __LOG_LEVELS__ = {
-        'debug': logging.DEBUG,
-        'info': logging.INFO,
-        'warn': logging.WARN,
-        'fatal': logging.FATAL,
-    }
+    logging.info("Starting to parse ...")
+    data = CreateParser().parse(open(filename).read(), file_name=filename)
+    logging.info("Parse completed")
 
-    @click.command()
-    @click.option(
-        '--log-level',
-        default='INFO',
-        type=click.Choice(list(__LOG_LEVELS__.keys()), case_sensitive=False),
-        help='Determines the verbosity of script output.')
-    @click.argument('filename')
-    def main(log_level, filename=None):
-        logging.basicConfig(
-            level=__LOG_LEVELS__[log_level],
-            format='%(asctime)s %(levelname)-7s %(message)s',
-        )
-
-        logging.info("Starting to parse ...")
-        data = CreateParser().parse(open(filename).read(), file_name=filename)
-        logging.info("Parse completed")
-
-        logging.info("Data:")
-        pprint.pp(data)
-
-    main(auto_envvar_prefix='CHIP')
+    logging.info("Data:")
+    pprint.pp(data)
