+++ scripts/py_matter_idl/setup.cfg
@@ -24,6 +24,13 @@ install_requires =
     lark
     jinja2
 
+[options.entry_points]
+console_scripts =
+    matter-data-model-xml-parser = matter.idl.data_model_xml:main
+    matter-idl-lint-parser = matter.idl.lint:main
+    matter-idl-parser = matter.idl.matter_idl_parser:main
+    matter-zapxml-parser = matter.idl.zapxml:main
+
 [options.packages.find]
 include = matter.idl*
